xml/parsing/manakai/attlists-3.dat

#data
<!DOCTYPE a[
  <!ATTLIST a b CDATA #FIXED>
]><a/>
#errors
1;1;no XML decl;;;s
2;15;default value not provided;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () FIXED ""
| <{} a>
|   b=""

#data
<!DOCTYPE a[
  <!ATTLIST a b CDATA #FIXED >
]><a/>
#errors
1;1;no XML decl;;;s
2;15;default value not provided;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT a >
|     b CDATA () FIXED ""
| <{} a>
|   b=""

#data
<!DOCTYPE x [<!ATTLIST hoge b (a|b)>]><x/>
#errors
1;1;no XML decl;;;s
1;14;empty attlist decl;;hoge;w
1;36;no attr default;;;m
#document
| <!DOCTYPE x>
|   <!ELEMENT hoge >
| <{} x>

#data
<!DOCTYPE x [<!ATTLIST hoge b (a|b) >]><x/>
#errors
1;1;no XML decl;;;s
1;14;empty attlist decl;;hoge;w
1;37;no attr default;;;m
#document
| <!DOCTYPE x>
|   <!ELEMENT hoge >
| <{} x>

#data
<!DOCTYPE x [<!ATTLIST hoge b (a|b)x>]><x/>
#errors
1;1;no XML decl;;;s
1;14;empty attlist decl;;hoge;w
1;36;no space between params;;;m
1;36;unquoted attr value;;;m
#document
| <!DOCTYPE x>
|   <!ELEMENT hoge >
| <{} x>

#data
<!DOCTYPE x [<!ATTLIST hoge b (a|b) x>]><x/>
#errors
1;1;no XML decl;;;s
1;14;empty attlist decl;;hoge;w
1;37;unquoted attr value;;;m
#document
| <!DOCTYPE x>
|   <!ELEMENT hoge >
| <{} x>
